@@ -20,9 +20,9 @@
     "start": "webpack-dev-server --progress"
   },
   "devDependencies": {
-    "@types/chai": "3.5.1",
+    "@types/chai": "3.5.2",
     "@types/mocha": "2.2.41",
-    "@types/node": "7.0.13",
+    "@types/node": "7.0.14",
     "chai": "3.5.0",
     "karma": "1.6.0",
     "karma-chrome-launcher": "2.0.0",
@@ -37,14 +37,14 @@
     "ncp": "2.0.0",
     "rimraf": "2.6.1",
     "ts-loader": "2.0.3",
-    "typescript": "2.2.2",
+    "typescript": "2.3.2",
     "webpack": "2.4.1",
-    "webpack-dev-server": "2.4.4"
+    "webpack-dev-server": "2.4.5"
   },
   "dependencies": {
-    "dashjs": "2.4.1",
+    "dashjs": "2.5.0",
     "eventemitter3": "2.0.3",
-    "hls.js": "0.7.5",
+    "hls.js": "0.7.6",
     "url-toolkit": "2.1.1",
     "webworkify": "1.4.0"
   },
